filemtime check added.
Installed via composer so code is different from what is in git.

My changes check the modified-timestamp of each file being added and compares it against the modified-timestamp of the cached min file. If there are any new changes then rebuild the cache file. Removed the unique id from the filename so only one cache file is created for every combination of files.

This removes the need to remove the old cache after every minor change.